requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
slowapi